Time: 0 ms (100%), Space: 27.6 MB (77.43%) - LeetHub

+++ 3300-minimum-element-after-replacement-with-digit-sum/3300-minimum-element-after-replacement-with-digit-sum.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int minElement(vector<int>& nums) {
-        int minim = 100;
+        int8_t minim = 100;
         for(int n : nums) {
-            int sum = 0;
+            int8_t sum = 0;
             while(n>0){
                 sum += n%10;
                 n /= 10;
